remove duplicate cross_entropy
